// src/app/shared/generated/model/mimeType.ts
/**
 * onecx-theme-bff
 * Backend-For-Frontend (BFF) service for onecx-theme. This API provides endpoints to manage Themes and related Images
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum MimeType {
    XIcon = 'image/x-icon',
    Svgxml = 'image/svg+xml',
    Png = 'image/png',
    Jpg = 'image/jpg',
    Jpeg = 'image/jpeg'
}